Timeout and context cancellation
This PR introduces support for timeout and context cancellation.
Before the execution of every new step the scheduler checks if the
context got cancelled or it is in timeout.

In this way we have a good way to terminate an execution or to track if
a plan did not end well because it was gonna be executed forever.

It is a good practice to always have a timeout for every execution
otherwise they can stay there looping forever.

Signed-off-by: Gianluca Arbezzano <gianarb92@gmail.com>